cleanup: move copy mc data into own handler struct
